Add labels to container image
Saves the values of the version, branch and revison build args as
labels.  Labels can be seen with podman-inspect, or in the webui for
quay.io. This is to make it easier to identify an image if, e.g., it was
pulled by digest or only has a generic tag like staging.

Label names are from https://github.com/opencontainers/image-spec/blob/main/annotations.md